Clean quick item input implementation
